GEODE-364: CI failure in SingleHopStatsDUnitTest.testClientStatsPR
This failure could also happen with the other test in this class:  a
client does put() operations that force the scheduling of a fetch of
PR metadata.  The client asserts that the metadata has been fetched,
but since the fetching is asynchronous it's possible for the client
to complete its put() operations and execute the assertion before the
asynchronous operation has completed.

The fix adds a WaitCriterion for the assertion.