Converting FilterProfile to use CopyOnWrite collections
There was a lot of overly-complicated code in FilterProfile to implement its own
copy-on-write but we now have CopyOnWriteHashSet and CopyOnWriteHashMap.  This
change-set converts FilterProfile to use those classes.

This change-set also includes refactoring of unregisterClientInterest performed in a
group refactoring session last week.  It adds unit tests for unregisterClientInterest
that give 96% code coverage for that algorithm and substantially increased coverage
for registerClientInterest as well.