GEODE-814: unit test DisconnectingOutOfOfHeapMemoryListener
Unit test has been added.

InternalDistributedSystem is no longer final so that unit tests can mock it.
The reason it was made final was because in an older version of the product
this class did license checking and we wanted to prevent it from being subclassed.
But now that it no longer does a license check the final was not needed.

Also the code coverage of OffHeapStorage was improved.