GEM-335: perf099 is more than 3x slower with new GMS
This lowers the membership event collection interval from 500ms to 300.  GemFire
8.2 used a 150ms interval so we might want to experiment with lowering it to
that in regression tests.

The wait period after sending a Leave message has been removed.  The
DistributionManager ShutdownMessage is already causing a Leave event to be
scheduled for a member when it is shutting down, so the Leave message isn't
needed unless the membership-manager is running outside of Geode/GemFire.

Finally, there was a race condition between the attemptToJoin method and
the installView method.  After sending a Join request it was possible for
a new view to be received and installed before attemptToJoin started its
wait() for a response.  This caused it to wait a full 12 seconds before
waking up and finding that isJoined had been set to true by the other thread.
Adding a check for isJoined before waiting for a response fixed this.