GEODE-864: Waiting for conflation thread in conflation tests
For parallel queues, there is a actually a separate conflation thread
that removes entries from the queue on conflation. The tests need to
wait for this thread to finish it's work.

As part of this change, I refactored the ParallelWANConflationDUnitTest
to remove a bunch of duplicate code, replaced the VM.invoke(String)
calls with lambdas, and scaled the test down to create fewer buckets and
do fewer puts so that it runs in 1/4 of the time.