GEODE-867: Undoing changes to pauseSender to wait for the pause
GEODE-864 changed WANTestBase.pauseSender to wait for the sender to
pause. But apparently in
ConcurrentSerialGatewaySenderOperationsDUnitTest the sender never
actually pauses, causing a hang in precheckin.

Temporarily undoing the changes for GEODE-864 and just changing the one
test to use pauseSenderAndWaitForDispatcher instead. That will leave
GEODE-864 fixed but also fix the hang until we can figure out why the
sender is not paused in this case.