GEODE-872: Removing duplicate code from ConcurrentSerialGatewaySenderOperationsDUnitTest
This test was almost an exact copy of
SerialGatewaySenderOperationsDUnitTest, with a few lines changes. The
non concurrent test had already fixed the suspect string, but the
concurrent test did not. By throwing away the code from Concurrent* and
making it extend the non-concurrent test, this bug is fixed.

The test has also been tuned to do fewer operations and take less time.