GEODE-871: build-up of sockets in TIME_WAIT on locator machine
This change-set alters the client to abort its TCP/IP connection to
the locator by enabling SO_LINGER and setting the timeout to zero
before it closes the connection.  The Locator closes its connection
first, which puts it into TIME_WAIT.  The client then aborts the
connection, which cleans up the Locator's TIME_WAIT connection.