GEODE-560: ClientServerTransactionCCEDUnitTest.testEventTracker
This test hasn't failed since a private build on 6 nov 2015 but I have
modified the test to find and install the transaction context in the test's
thread if it finds that the verification task is running in a different
RMI thread.

I've left an error message in that code so that we can tell that the fix
worked - it should fail with a suspect string but otherwise pass.  If this
happens the error message can be removed from the test.