GEODE-841: ClientMembershipDUnitTest.testGetNotifiedClients failed
This test threw an exception when it received a client identifier with a
numeric IP address instead of a host name.  On the servers these IDs are
deserialized and a canonical host name is picked up from SocketCreator,
but this is accessing a static cache of host names that may have been
initialized during a test that turned of DNS name lookup altogether in
order to avoid network delays and possible hangs.

The fix is to just clear the cache between tests.