GEODE-481: CI failure: QueryUsingFunctionContextDUnitTest.testQueriesWithFilterKeysOnPRWithBucketDestroy
The error handling for functions is different depending on if it failed locally or remotely.
In this case, the function was failing on the remote node and the try/catch block would have it send
the exception, however there are two types of send exception calls.
We needed it to use the result sender's way to send the exception (possibly less header data)