GEODE-550 CI failure: FixedPRSinglehopDUnitTest.test_MetadataContents
This test hasn't failed in over 3 months and that was on a private build
with no artifacts.  I've changed the test to include details about the
end state of the client's metadata so that if this test fails again we
will know more about why it failed.

I'm closing the ticket so that CI monitors will reopen it and not just
report that it failed again in an email.