GEODE-323: Primary Key indexes are not being properly used for Partitioned Region Queries
Iterator naming is set at the bucket index level, however lookups are done on the PartitionIndex level.
When a lookup occurs, it will not find the correct mapping to index iterator name. This leads to no
matching indexes when expecting a primary key index lookup.
The fix is to populate a shared map at the PartitionIndex level for all bucket indexes and do the look
up at the PartitionIndex level.

Fix bug where isPopulate flag is not being set on PrimaryKey indexes that prevented them from being used.