Removing statics from the lambda expressions
A bunch of tests were failing because the location in which static
fields are evaluated has changed. Before, a static in the parameter list
was evaluated on the controller. Now it is evaluated on the remote VM.

Converting the statics to fields or local variables so they are captured
and serialized.