Porting a fix for a race condition to develop
This bugfix was put into a GemFire maintenance branch last summer but never
made it to the develop branch.  Someone ran into the problem again recently
on a Geode develop build.

If a client retries a bulk operation (putAll / removeAll) the second attempt
can encounter the original attempt and cause the second attempt to fail as
it cannot find the version tag for the original operation in the region's
event tracker.  This change records the version tag before information about
the bulk-op is recorded, so we know that the version tag will be available
for the reattempted operation.