ensure that rejected messages propagate an exception
MessageJUnitTest has a test to ensure that large messages cause an exception
to be thrown during transmission, but it does not ensure that the exception
is propagated up through the OpExcecutor.  This new test ensures that this
happens.

I also renamed a couple of testBug.... methods to have more meaningful names