adding unit tests for REST API
A number of unit tests were still sitting in the work queue to be moved to
open-source including all of the REST API tests.  This change set shows the
build changes and AgentUtil changes needed to accomodate the new tests but
does not show the new tests.  The tests are being added to the
geode-assembly project, and all of the support classes in geode-core's
com.gemstone.gemfire.rest.internal.web.controllers package are being moved
to that project.

The tests will not run in Eclipse since a WAR file is not being created by
the Eclipse projects.  I've opened a JIRA ticket about that.