GEODE-27: Enabling transitive dependencies in the build
Removing the code from the build that disabled transitive dependencies.
Adding excludes and forcing some versions to clean up the list of jars
we are pulling in to match what it was before I enabled transitive
dependencies.

A couple of versions don't match because the versions we were pulling in
before were not consistent across submodules for commons logging and
commons-beanutils.

Some of these excludes, and all of the forced versions look suspect.
They need to be reviewed to see if we really need them.