GEODE-27: A couple of tools to dump dependencies
Adding a task to dump all jars in the binary distribution

This makes it easy to see what dependencies we actually end up shipping
with the product. This dumps all jars found in the distribution or
embedded within war files in the distribution.

Adding a task to determine if a dependency is used

This task checks to see if a dependency is used in the source code.