GEODE-952: Some time many unit tests fails because locator was ForcedDisconnect
We're seeing a number of similar failures that all seem to be caused by JVMs
pausing and being kicked out of the distributed system.   This change-set
enables creation of a heap dump if a member is forced out of the system and
JVM pauses have been detected.  This will give us artifacts that we can
analyze to help determine what's going on.