GEODE-1044: InitialImagePut executes BEFORE_UPDATE_OP on indexes for updates
When entries are recovered from persistence, the values are added to the index,
however a gii can then occur where the initialImagePut needs to properly
alert the index that an update is occuring.  The BEFORE_UPDATE_OP allows the index
to retreive the oldKey and notifies the index that an update is occuring.

Refactor QueryTestUtils and removed unused code
Minor modification for QueryTestUtils to be more generic and not be specific to Portfolios