GEODE-27: Marking jars as optional in the generated pom
Adding support to mark jars as optional by setting an ext.optional
property in the dependency declaration.

Adding the optional property to jars in the geode-core project that are
not required for geode to work.

Some of these optional dependencies will go away if we can move things
like the REST API, gfsh, and management classes into another subproject.