GEODE-934: CI failure: DurableClientSimpleDUnitTest.testReadyForEventsNotCalledImplicitlyWithCacheXML.
In this test the client sees previously delivered events, when it disconnects and reconnects immediately.
This happens when the message dispatcher is in the process of removing the messages while its in the pause state; and client reconnects at the same time. When client reconnects, it clears out the acked-event list, this causes the message dispatcher thread not to remove dispatched event from the queue. The fix is to take the pause lock and start removing the dispatched event, this blocks the client reconnect process to wait tille the messages are removed.

Testing:
The test used to fail once in 4 runs. With the change it run for 50 times.
precheckin