GEODE-947 Client log message misleading
Logging of local-mode or client-mode is now done in GemFireCacheImpl when
the cache is created.  If there is a connection pool or the cache is marked
isClient we want to say it's running in client-mode.  If it's not in client-
mode and the DistributedSystem is not actually configured to be able to
distribute anything we say it's running in local-mode.

I corrected the text of the local-mode message to remove the reference to
multicast since multicast discovery is not supported in Geode.