GEODE-956 Product use log shows duplicate entries
The product use logger was adding up the load of all of the servers and
displaying it, but each load is a float between zero and 1.0 so the total
was always zero.  I've modified it to figure out the correct connection
count for each CacheServer and also log the subscription queue count.

The list of servers now has duplicate entries removed by putting all of
the servers into a Set.