GEODE-988: CI Failure: CqPerfUsingPoolDUnitTest.testMatchingCqs failed with AssertionFailedError
Test issue. The test is not waiting for events to arrive for all the CqListeners; it waits for one of the CqListener (thae last one registered) and validates for all CqListener; based on the order at which CQlistners are invoked (and thread scheduling), when check for expected events are made, the events may not have arrived at all CqListeners.

Added wait logic in the Listner code, so that we wait for all the expected event to arrive.