GEODE-1123: Exclude the locator in StreamingOperationManyDUnitTest
This test was calling getOtherDistributionManagerIds, which included
that locator. There seemed to be some race where the test would pass and
find 4 but if I added a sleep the test would always fail with 5 members
(the 4 VMs and the locator).

The test now doesn't use getOtherDistributionManagerIds at all, since it
really was just trying to get the list of member ids of the other VMs.