GEODE-1031 Now pool closes queueMge before connectionMgr
At cache close, there is possibility that queueMgr is trying to
recover queue but cache/pool trying to close cache with keep-alive
false flag. During this time there is possibility that client(queueMgr)
may be unable to close socket properly as it doesn't know about
keep-alive flag. Thats why closing connectionMge after queueMgr
so that it should be able to send keep-alive flage properly
to server.