GEODE-1025: Removing bundled SDG and special handling of spring jars
Spring Data Gemfire will no longer be bundled with geode, instead users
will be expected to manually add it to the classpath if they want to
start with spring.

Removing the spring jars from the lib directory, and the special code
that automatically adds the spring jars to the classpath when starting
with --spring-xml-location.