GEODE-704: Added check/wait for interest re-registration to complete after cache-server is restarted.
CI failure: HAInterestPart2DUnitTest.testRefreshEntriesFromPrimaryWhenDSMDetectsServerLive

In the test client re-registers the interests after cache-server is started and expects
entries to be refreshed (it stops and starts the cache-server, without closing the cache)...
It looks like we are not waiting enough to see the client registration is complete
and events are refreshed (the re-registratin happens with redundancy satisfier thread).

Added wait/check to see interest registration is complete before check is done for the
cache entries...

Testing:
Run the test 100+ times.