GEODE-1187 a server launched by Gfsh deletes PID file during auto-reconnect
ServerLauncher needed to be made aware of auto-reconnect.  I changed
getCache() to look for a new, reconnected cache and I changed isWaiting()
to see if the current cache is in a reconnecting state.

I also changed the shutdown logic to cancel reconnect attempts if the
cache is in that state when a stop is requested.