GEODE-1025: Force the version of spring core we use
Spring shell depends on an old version of spring-core. Once all of the
rest of the spring dependencies were removed from geode-core, the spring
shell dependency won and caused us to bundle the older version of spring
core in the lib directory.

With this change, we all always use the desired version of spring-core.