GEODE-1226: Prompt for a password to sign and upload to maven
The signArchives and uploadArchives tasks require passwords to work.
Adding a flag to request a password prompt to fill in a password.

./gradlew uploadArchives -Paskpass

This will pop up a swing dialog box asking for passwords to sign and
upload archives.

This is not on by default because we don't want developers to have to
enter a password for every build. The build target depends on
signArchives but normally will skip it if the password property is not
set.