GEODE-1174 CI failure: UniversalMembershipListenerAdapterDUnitTest.testSystemClientEventsInServer
When a client closes its connection Pool a CloseConnection request is sent to
the server and then sockets are closed.  Unfortunately this sometimes results
in the socket being closed before the operation reaches the server and the
server sees the client as having crashed.  This is mostly a problem if the
client is sending keepAlive==true but it is also what is causing this test
to fail intermittently.

I've changed CloseConnectionOp to wait for a reply and have changed the
server to send this reply.  It's possible for the client to get an
EOFException instead of a reply, but at least this causes the client to
wait for the server to terminate the connection so that we know that the
server has received the request and acted on it.

Since old GemFire clients might be used with Geode I've added checks for them
and avoid sending a reply in that case.