GEODE-1237 remove the check for network-partition-detection uniformity
Each member now determines whether to force shutdown due to loss of quorum.
They also decide whether locators are preferred as membership coordinators.
This primarily meant removing the method GMSJoinLeave
.inheritSettingsFromLocator() but the ID settings needed to be made in
other places (JGroupsMessenger.establishLocalAddress() and
GMSJoinLeave.started()).

I also noticed that some of the code was still using "splitBrain" and
"SB" instead of networkPartitionDetection and "NPD" and fixed that.

One test had to be removed from LocatorDUnitTest and a couple of others
needed to be adjusted to have all of the members set network-partition
-detection-enabled=true.  Precheckin is running and I'll fix any other
tests that have this problem.