GEODE-357: Remove third server from testInitiateFailoverByCacheOperationThreads_Secondary
This test was intermittently connecting to either 2 servers or 3. If the
test never connected to the third server, it would fail waiting for the
live server count to be 2.

Also cleaned up some code to use Awaitility and to have a lower ping
interval, because the ping is what causes us to disconnect from the
stopped server.