GEODE-1391 JsonFormatter javadocs need to be cleaned up
I've made another pass at JSONFormatter javadocs, putting the JSON <-> Java
types in a table and correcting a few grammatical errors.  I also found that
the class was being referred to as PdxFormatter in a lot of places and that
its test was also using this name instead of JSONFormatter.