GEODE-1415 Disable logging of banners in distributedTests
Banners are normally logged during creation of an InternalDistributedSystem.
In DUnit tests the output is filled with these banners and they all have
the same information, so they are just clutter.

This change-set disables banners in non-locator VMs launched by
DUnitLauncher/ProcessManager.  Precheckin testing shows that the
amount of output for geode-core/distributedTests alone used to be over 1gb
and is now less than 500mb.