GEODE-1456: fix race in GemFireCacheImplTest
The race was caused by the GemFireCacheImpl constructor creating a real TypeRegistry.
The TypeRegistry creates a region which scheduled an async create region event with the
event pool. So when the test set "initialCount" this create region event might not have
completed so the initialCount would be zero. The test then scheduled MAX_THREADS tasks
and waits until that many complete. But the create region event could also complete causing
getCompletedTaskCount() to return a value 1 more than the test expected which caused it
to intermittently timeout.

The test now mocks the TypeRegistry so the only events scheduled with the pool come from
the unit test.