GEODE-1494: Changing stats list to be a CopyOnWriteArrayList
The statsList used to track Statistics in InternalDistributedSystem was
managed under a lock. That was problematic, because code that iterates
on the list needs to hold the lock. If the code invoking callbacks for
GEODE-1494 was slow, that means the lock would block other stats
operations.

Changing this list to a copy on write array list to holding a lock for a
long period of time.