GEODE-1422: Removing check to skip enqueuing temp events
This check was performed outside of synchronizing on the lock, which
means that some events could be added to the temp events after the
check. This caused the test to fail due to temp events being left in the
map.

Also, fixing two race conditions in testParallelPropagationSenderStartAfterStop_Scenario2

1) It did some puts and stopped the sender without waiting for the
puts to complete. So they may not be available on the remote side

2) It started some puts asynchronously, and then started the sender
later. The puts performed before the sender started may not be present
on the remote side.