Merge #158 into feature/GEODE-33 GEODE-1530: geode-examples setEnv.sh script needs to export path
Without an export of the path, using the GEODE_HOME env variable,
the scripts/startAll.sh script will pick up the first gfsh in the
path that it finds.

Also updated the scripts/stopAll.sh script to run the setEnv.sh
script, so that stopAll.sh uses the correct gfsh.