GEODE-1544: Fail to initiate final check when network partition occurs
Currently, suspect request is only sent to a limited number of members,
when there are more than 4 members in the distributed system. It is
possible that the recipients of suspect request is unreachable due to
network partition. This fix makes sure that the recipients of suspect
request are not the suspect members. So that suspect request will be
processed.