GEODE-11: Pagination of Lucene Query Results after entries are destroyed from the region
If a region entry is destroyed after a query is executed, the pages.next() call returns
a null value for the deleted entry. The code is changed to ignore deleted entries and
fetch the next available entry so that the page count remains consistent. Added tests to
verify scenarios of single, multiple and all region entry deletions.

Added a test to verify that all entries are fetched in one page when pagination is
disabled.

Signed-off-by: Anilkumar Gingade <agingade@pivotal.io>

This closes #178