GEODE-11: Implementing GFSH command for listing lucene indexes.
Added the LuceneIndexCommands class that implements the list lucene index functionality in GFSH and the LuceneListIndexFunction class
that gets the information about lucene indexes from the cache. Also added a test to verify this functionality. The
org.springframework.shell.core.CommandMarker file is required to register the lucene commands in GFSH.

Signed-off-by: Gester Zhou <gzhou@pivotal.io>