GEODE-11 Adding tests of closing a cache during index updates
Testing failover of lucene indexes by closing the cache while in the
middle of updating lucene indexes. Currently there are tests for closing
the cache before the index repository commit, and also during the commit
after a fixed number of updates to the underling index data regions.

I refactored the lucene tests to use 7 buckets, rather than 113, so they
take less time and are easier to debug.

I also removed a call to Thread.interrupt in the WAN code because it was
interrupting itself in my callback. We should never be using interrupt
in the product.