GEODE-1669: Made changes to add/remove list of interested keys using addAll() and removeAll()
The client interests are managed in "FilterProfile" class on server side. These are maintained
using the concurrent data structures CopyOnWriteHashSet and CopyOnWriteHashMap...

When set of keys are registered from client, the keys are added to CopyOnWriteHashSet one by
one (FilterProfile.registerClientInterestList()); Where a new HashSet is created each time when
an entry is added, which could impact performance based on the number of keys registered.

Testing:
precheckin
Added new unit test.