GEODE-1583: Pass the keep alive flag when destroying a connection
The queue manager was closing connections without sending the
CloseConnectionOp to the server. Changing the code to call close instead
of destroy and passing the keepAlive flag along with the close. Also,
changing the cache to set keepAlive *before* marking the cache as
closed.

This should prevent any races where cache closure triggers the cache
client updater to destroy connections or close them with the wrong
keepAlive setting while the cache is being closed.