GEODE-1700: fix stat corrupting when async queue full
A common method is now used for writing async ops to disk.
The same method is now used for both flushing an item that
has been taken out of the queue and for when the queue is
full causing a sync write to be done.

Also did a minor refactoring to get rid of the "removed" flag.
This revision fixes a bug introduced in GEODE-93 and should be
included with that fix.