GEODE-1714: fix disk stats
A number of places were found that did not correctly
change the DiskRegionStats: entriesInVM and entriesOnlyOnDisk.
Also found places that did not correctly change
the entry count and size of a PR bucket region.
EntriesInVM is supposed to only count entries that have a value
stored in the JVM. But it was also counting invalid entries.
Since invalid entries do not have a value this stat no longer counts them.