GEODE-1687: Added null check for CQs proxy connection.
When a connection relating to CQ is established or dropped, CqStatusListner is invoked
to indicate connection/subscription status. This allows application to handle if all
the subscription connection is lost for CQs.

Looking at the code path for CqStatusListener there is a chance of NPE getting thrown
in CqServiceImpl.invokeCqsConnected when CQ is executed using the Bridge loader
interface (old way of client-server connection)...