GEODE-1728: Removing the dispatcher elements from the session filter
The geode session filter is being applied to all requests. Because of
that, all request objects should already be wrapped in a RequestWrapper.
By adding these dispatcher elements to the filter, the filter was being
applied multiple times to the request when a request is forwarded or
included using the RequestDipatcher. That is unnecessary, and
problematic if we can't determine that the request has already been
wrapped.